--- a/src/Parlot/Fluent/OneOf.cs
+++ b/src/Parlot/Fluent/OneOf.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Parlot.Fluent
+﻿namespace Parlot.Fluent
 {
     /// <summary>
     /// OneOf the inner choices when not all parsers return the same type.
@@ -9,16 +6,18 @@ namespace Parlot.Fluent
     /// </summary>
     public sealed class OneOf : Parser<ParseResult<object>>
     {
-        public OneOf(IList<IParser> parsers)
+        private readonly IParser[] _parsers;
+
+        public OneOf(IParser[] parsers)
         {
-            Parsers = parsers ?? throw new ArgumentNullException(nameof(parsers));
+            _parsers = parsers;
         }
 
-        public IList<IParser> Parsers { get; }
+        public IParser[] Parsers => _parsers;
 
         public override bool Parse(ParseContext context, ref ParseResult<ParseResult<object>> result)
         {
-            if (Parsers.Count == 0)
+            if (_parsers.Length == 0)
             {
                 return false;
             }
@@ -27,9 +26,9 @@ namespace Parlot.Fluent
 
             var start = context.Scanner.Cursor.Position;
 
-            for (var i = 0; i < Parsers.Count; i++)
+            foreach (var parser in _parsers)
             {
-                if (Parsers[i].Parse(context, ref parsed))
+                if (parser.Parse(context, ref parsed))
                 {
                     result.Set(parsed.Buffer, parsed.Start, parsed.End, parsed);
                     return true;
@@ -51,26 +50,29 @@ namespace Parlot.Fluent
     /// <typeparam name="T"></typeparam>
     public sealed class OneOf<T> : Parser<T>
     {
-        public OneOf(IList<IParser<T>> parsers)
+        private readonly IParser<T>[] _parsers;
+
+        public OneOf(IParser<T>[] parsers)
         {
-            Parsers = parsers;
+            _parsers = parsers;
         }
-        public IList<IParser<T>> Parsers { get; }
+
+        public IParser<T>[] Parsers => _parsers;
 
         public override bool Parse(ParseContext context, ref ParseResult<T> result)
         {
             context.EnterParser(this);
 
-            if (Parsers.Count == 0)
+            if (Parsers.Length == 0)
             {
                 return false;
             }
 
             var start = context.Scanner.Cursor.Position;
 
-            for (var i = 0; i < Parsers.Count; i++)
+            foreach (var parser in _parsers)
             {
-                if (Parsers[i].Parse(context, ref result))
+                if (parser.Parse(context, ref result))
                 {
                     return true;
                 }