--- a/src/Parlot/Fluent/ParseResult.cs
+++ b/src/Parlot/Fluent/ParseResult.cs
@@ -1,38 +1,32 @@
-﻿using System;
-
-namespace Parlot
+﻿namespace Parlot
 {
     public struct ParseResult<T>
     {
-        public ParseResult(string buffer, TextPosition start, TextPosition end, T value)
+        public ParseResult(string buffer, in TextPosition start, in TextPosition end, T value)
         {
             Buffer = buffer;
             Start = start;
             End = end;
-            Length = end - start;
             Value = value;
         }
 
-        public void Set(string buffer, TextPosition start, TextPosition end, T value)
+        public void Set(string buffer, in TextPosition start, in TextPosition end, T value)
         {
             Buffer = buffer;
             Start = start;
             End = end;
-            Length = end - start;
             Value = value;
         }
 
-        public TextPosition Start { get; private set; }
-
-        public TextPosition End { get; private set; }
-
-        public int Length { get; private set; }
-        public string Buffer { get; private set; }
-
-        public string Text => Buffer?.Substring(Start.Offset, Length);
-
-        public ReadOnlySpan<char> Span => Buffer.AsSpan(Start.Offset, Length);
-
-        public T Value { get; private set; }
+        public TextPosition Start;
+        public TextPosition End;
+        public string Buffer;
+        public T Value;
+    }
+    
+    // if really needed, allows less fields if can be computed, no need to methods
+    internal static class ParseResultExtensions
+    {
+        public static int GetLength<T>(this ParseResult<T> result) => result.End - result.Start;
     }
 }