--- a/src/Parlot/ThrowHelper.cs
+++ b/src/Parlot/ThrowHelper.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace Parlot
+{
+    using System.Diagnostics.CodeAnalysis;
+
+    internal static class ThrowHelper
+    {
+        [DoesNotReturn]
+        public static void ThrowArgumentNullException(string paramName)
+        {
+            throw new ArgumentNullException(paramName);
+        }
+    }
+}