Wanted so see what's this new shiny thing a saw things that might improve perf, so here's some.

* improve array access by having array as local using uint access to avoid bound checks  when needed / using foreach if possible
* prefer read-only structs, get/set properties and methods are bad for structs as they cause copying for CLR to be safe
   * generally should make structs less class-like, it's kind of either or with the properties and methods, extension methods can help with structs
* tried to make some property access inlined

To further improve performance, some suggestions:

* Ditch the `IParser`/`IParser<T>` and always use abstract class access, interface dispatch is more costly, these are called a lot
* Consider if possible for benchmarks to use TextSpan instead of string for getting JSON properties, it might make the parsers con-comparable though, the `ToString()` needed for materializing properties isn't awfully nice. I did try this with custom Marvin hasher for `ROS` (`netstandard2.1` doesn't have `GetHashCode` for `ROS`)
* traversing the string with pointers instead of index might be faster, haven't tried it out though

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 7 2700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.101
  [Host]   : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  ShortRun : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```

## Parlot.Benchmarks.ExprBench

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParlotRawSmall|784.8 ns|308.0 ns|0.1411|0.0000|0.0000|592 B|
| **New** |	| **759.9 ns (-3%)** | **305.91 ns** | **0.1411 (0%)** | **0.0000** | **0.0000** | **592 B (0%)** |
| Old |ParlotFluentSmall|2,623.9 ns|320.7 ns|0.1755|0.0000|0.0000|744 B|
| **New** |	| **1,975.2 ns (-25%)** | **54.83 ns** | **0.1755 (0%)** | **0.0000** | **0.0000** | **736 B (-1%)** |
| Old |PidginSmall|14,292.6 ns|376.3 ns|0.1831|0.0000|0.0000|816 B|
| **New** |	| **14,645.4 ns (+2%)** | **1,034.18 ns** | **0.1831 (0%)** | **0.0000** | **0.0000** | **816 B (0%)** |
| Old |ParlotRawBig|3,909.5 ns|1,367.0 ns|0.6447|0.0000|0.0000|2712 B|
| **New** |	| **4,059.4 ns (+4%)** | **106.52 ns** | **0.6485 (+1%)** | **0.0000** | **0.0000** | **2712 B (0%)** |
| Old |ParlotFluentBig|13,842.5 ns|2,039.8 ns|0.7477|0.0000|0.0000|3136 B|
| **New** |	| **10,241.6 ns (-26%)** | **599.34 ns** | **0.7477 (0%)** | **0.0000** | **0.0000** | **3128 B (0%)** |
| Old |PidginBig|72,480.7 ns|1,090.6 ns|0.8545|0.0000|0.0000|4072 B|
| **New** |	| **72,651.2 ns (0%)** | **2,222.91 ns** | **0.8545 (0%)** | **0.0000** | **0.0000** | **4072 B (0%)** |


## Parlot.Benchmarks.JsonBench

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |BigJson_Parlot|367.6 μs|2.37 μs|24.9023|3.4180|0.0000|101.84 KB|
| **New** |	| **292.0 μs (-21%)** | **3.01 μs** | **24.9023 (0%)** | **3.4180 (0%)** | **0.0000** | **101.84 KB (0%)** |
| Old |BigJson_Pidgin|556.1 μs|6.77 μs|24.4141|2.9297|0.0000|101.7 KB|
| **New** |	| **553.3 μs (-1%)** | **3.96 μs** | **24.4141 (0%)** | **2.9297 (0%)** | **0.0000** | **101.7 KB (0%)** |
| Old |BigJson_Sprache|3,595.3 μs|254.49 μs|1308.5938|50.7813|0.0000|5349.63 KB|
| **New** |	| **3,511.6 μs (-2%)** | **39.13 μs** | **1308.5938 (0%)** | **50.7813 (0%)** | **0.0000** | **5349.63 KB (0%)** |
| Old |BigJson_Superpower|2,133.4 μs|148.34 μs|222.6563|3.9063|0.0000|913.43 KB|
| **New** |	| **2,144.2 μs (+1%)** | **126.62 μs** | **222.6563 (0%)** | **3.9063 (0%)** | **0.0000** | **913.43 KB (0%)** |
| Old |LongJson_Parlot|289.6 μs|2.72 μs|25.3906|0.4883|0.0000|104.39 KB|
| **New** |	| **236.2 μs (-18%)** | **1.35 μs** | **25.3906 (0%)** | **0.2441 (-50%)** | **0.0000** | **104.38 KB (0%)** |
| Old |LongJson_Pidgin|491.8 μs|12.61 μs|25.3906|3.9063|0.0000|104.25 KB|
| **New** |	| **485.2 μs (-1%)** | **4.95 μs** | **25.3906 (0%)** | **2.4414 (-38%)** | **0.0000** | **104.25 KB (0%)** |
| Old |LongJson_Sprache|2,847.4 μs|161.38 μs|1054.6875|11.7188|0.0000|4311.36 KB|
| **New** |	| **2,895.6 μs (+2%)** | **161.87 μs** | **1054.6875 (0%)** | **11.7188 (0%)** | **0.0000** | **4311.36 KB (0%)** |
| Old |LongJson_Superpower|1,734.5 μs|34.44 μs|171.8750|3.9063|0.0000|706.79 KB|
| **New** |	| **1,730.9 μs (0%)** | **103.56 μs** | **171.8750 (0%)** | **3.9063 (0%)** | **0.0000** | **706.79 KB (0%)** |
| Old |DeepJson_Parlot|248.6 μs|35.45 μs|20.0195|0.4883|0.0000|82.39 KB|
| **New** |	| **198.4 μs (-20%)** | **3.77 μs** | **20.0195 (0%)** | **0.2441 (-50%)** | **0.0000** | **82.38 KB (0%)** |
| Old |DeepJson_Pidgin|573.1 μs|49.33 μs|45.8984|0.9766|0.0000|187.79 KB|
| **New** |	| **567.8 μs (-1%)** | **20.04 μs** | **45.8984 (0%)** | **0.9766 (0%)** | **0.0000** | **187.79 KB (0%)** |
| Old |DeepJson_Sprache|3,190.5 μs|37.78 μs|562.5000|226.5625|0.0000|2946.56 KB|
| **New** |	| **3,211.3 μs (+1%)** | **261.43 μs** | **554.6875 (-1%)** | **218.7500 (-3%)** | **0.0000** | **2946.56 KB (0%)** |
| Old |WideJson_Parlot|204.9 μs|3.67 μs|11.7188|0.2441|0.0000|48.56 KB|
| **New** |	| **178.0 μs (-13%)** | **4.43 μs** | **11.7188 (0%)** | **0.7324 (+200%)** | **0.0000** | **48.55 KB (0%)** |
| Old |WideJson_Pidgin|290.6 μs|4.36 μs|11.7188|0.9766|0.0000|48.42 KB|
| **New** |	| **293.6 μs (+1%)** | **2.93 μs** | **11.7188 (0%)** | **0.9766 (0%)** | **0.0000** | **48.42 KB (0%)** |
| Old |WideJson_Sprache|1,676.3 μs|123.49 μs|683.5938|11.7188|0.0000|2797.28 KB|
| **New** |	| **1,635.4 μs (-2%)** | **113.14 μs** | **683.5938 (0%)** | **11.7188 (0%)** | **0.0000** | **2797.28 KB (0%)** |
| Old |WideJson_Superpower|1,108.5 μs|65.53 μs|111.3281|3.9063|0.0000|459.74 KB|
| **New** |	| **1,104.4 μs (0%)** | **14.52 μs** | **111.3281 (0%)** | **3.9063 (0%)** | **0.0000** | **459.74 KB (0%)** |