@sebastienros 2021-01-03
This PR doesn't have complex changes and yet produces great benefits. I wonder which one is the most significant. I might try a few of the most obvious ones to see.

For the Json benchmarks changes, you are right as I don't want to change it. I want to comply with what Pidgin was doing. The goal is just to compare the parsers, not the raw outcome of the benchmark.

The IParser usage is an issue right now. I have struggled finding a good compromise with the class heerarchy, and how to provide a good Fluent API. I think I haven't found the best approach. Maybe I should try to ditch the interface altogether if it doesn't provide any meaningful abstraction. Maybe the base abstract class would be sufficient. I would also love to have a covariant type that would allow for Parser<T> to be castable to Parser<object>.

Good idea for the string pointers. That's what the ROS algorithms do internally. Maybe the scanner/cursor should only deal with a ROS.